refactor: move and rename ShareCardGroup
