+++ src/components/TeacherComponents/LessonShareCardGroup/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./LessonShareCardGroup";
